refactored RenderingContext out
It doesn't make sense for the end-user to really have access to the RenderingContext - it's only necessary for SOME Graphics API (ex. OpenGL), and it needs to be handled carefully with regards to threading.

So I moved the API calls into a RenderingState class, and made Graphics able to access it cleanly.

TODO: OpenGL implementation needs to do some more thread checks on a bunch of its code